+++ kubejs/server_scripts/recipes/ars_nouveau/shaped.js
@@ -91,7 +91,5 @@ ServerEvents.recipes((event) => {
 
     recipes.forEach((recipe) => {
         event.shaped(recipe.output, recipe.pattern, recipe.key).id(recipe.id);
-
-        
     });
 });