+++ kubejs/server_scripts/recipes/arts_and_crafts/shaped.js
@@ -15,7 +15,5 @@ ServerEvents.recipes((event) => {
 
     recipes.forEach((recipe) => {
         event.shaped(recipe.output, recipe.pattern, recipe.key).id(recipe.id);
-
-        
     });
 });