+++ CHANGELOG.md
@@ -5,6 +5,7 @@ NeoForge-1.21.1-21.1.232 |
 #### ⭐ Improvements
 
 - Pink Slime output has been increased slightly [(\#249)](https://github.com/EnigmaticaModpacks/EnigmaticSkies/issues/249)
+- Add recipes for Nylium [(\#249)](https://github.com/EnigmaticaModpacks/EnigmaticSkies/issues/249)
 
 #### 🦟 Bugs Fixed